Create fw01.config.week9.txt
Note that this is just the random order that vyOS spit out the configuration commands, it's important to understand the dependencies when configuring the firewall, and it's good practice to network and configure all of the services before locking down the firewall with zones, since there's packages that need to be pulled from the internet, etc. There are also a few areas where the IPs are not used anymore, and they need to reflect the new host names and IP addresses.